Move the Payload struct and related methods to its own file.
